+++ web/website/themes/prql-theme/static/plugins/highlight/prql.js
@@ -38,9 +38,8 @@ formatting = function (hljs) {
     "_is_null",
     // Misc functions
     "from_text",
-    // String functions
-    "lower",
-    "upper",
+    // String functions module
+    "str",
     // Window functions
     "lag",
     "lead",